Rev 232 API changes

+++ src/main/java/com/stevenwaterman/blindfold/DisableRenderCallbacks.java
@@ -65,7 +65,7 @@ public class DisableRenderCallbacks implements DrawCallbacks
 	}
 
 	@Override
-	public boolean tileInFrustum(Scene scene, int pitchSin, int pitchCos, int yawSin, int yawCos, int cameraX, int cameraY, int cameraZ, int plane, int msx, int msy)
+	public boolean tileInFrustum(Scene scene, float pitchSin, float pitchCos, float yawSin, float yawCos, int cameraX, int cameraY, int cameraZ, int plane, int msx, int msy)
 	{
 		return false;
 	}
